Potentiometer
Using a potentiometer with 10k resistance. turning the knob changes the output voltage between the source (in this case 5V) and ground. if you turn it all the way to the left the output is 0v and if you turn it all the way right it's 5v. analogRead is used to read the voltage, but it returns a value between 0 and 1023, which is why you have to multiply the read value by (float)(5/1023) in order to get a value between 0 and 5.