Dimmable LED
Reading value from potentiometer, converting with (float)(255/1023). This is because the second parameter of analogRead() takes a voltage read between 0 and 255. Turning the knob to the left turns it off and turning it right will turn it fully on, the brightness in between changes according to value read from the pot val.