While Loop
Turns on "error" (red) led when the output voltage read (in analog) is over 1000. Inside the while loop, still checking and setting the potentiometer variable to the read value so that it  can escape the loop if the variable changes, otherwise it would loop forever. LED turns off once while loop is exited.